+++ Utilities/StringHelper.php
@@ -1,13 +1,13 @@
 <?php
 /**
  * @package      Prism
- * @subpackage   Strings
+ * @subpackage   Utilities
  * @author       Todor Iliev
  * @copyright    Copyright (C) 2015 Todor Iliev <todor@itprism.com>. All rights reserved.
  * @license      GNU General Public License version 3 or later; see LICENSE.txt
  */
 
-namespace Prism;
+namespace Prism\Utilities;
 
 use Joomla\Utilities\ArrayHelper;
 
@@ -18,37 +18,16 @@ defined('JPATH_PLATFORM') or die;
  * This class contains methods that are used for handling strings.
  *
  * @package     Prism
- * @subpackage  Strings
- *
- * @deprecated 1.10  Use {@link \Prism\Utilities\StringHelper} instead unless otherwise noted.
+ * @subpackage  Utilities
  */
-class String
+class StringHelper
 {
-    protected $content;
-
-    /**
-     * Initialize the object.
-     *
-     * <code>
-     * $content = "If you can dream it, you can do it."
-     *
-     * $string = new Prism\String($content);
-     * </code>
-     * 
-     * @param string $content
-     */
-    public function __construct($content = "")
-    {
-        $this->content = (string)$content;
-    }
-
     /**
      * The method generates random string.
      * You can set a prefix and specify the length of the string.
      *
      * <code>
-     * $hash = new Prism\String;
-     * $hash->generateRandomString(32, "GEN");
+     * $hash = Prism\Utilities\StringHelper::generateRandomString(32, "GEN");
      *
      * echo $hash;
      * </code>
@@ -57,23 +36,19 @@ class String
      * @param string  $prefix A prefix, which will be added at the beginning of the string.
      *
      * @return string
-     *
-     * @deprecated since v1.10
      */
-    public function generateRandomString($length = 10, $prefix = "")
+    public static function generateRandomString($length = 10, $prefix = '')
     {
         // Generate string
         $hash = md5(uniqid(time() + mt_rand(), true));
         $hash = substr($hash, 0, $length);
 
         // Add prefix
-        if (!empty($prefix)) {
+        if (trim($prefix)) {
             $hash = $prefix . $hash;
         }
 
-        $this->content = $hash;
-
-        return $this;
+        return $hash;
     }
 
     /**
@@ -89,25 +64,23 @@ class String
      *     "position" => 0 // 0 for symbol on the left side, 1 for symbole on the right side.
      * );
      *
-     * $amount = new Prism\String(100);
-     * $amount->getAmount(GBP, $options);
+     * $amount = Prism\Utilities\StringHelper::getAmount(100, GBP, $options);
      *
      * echo $amount;
      * </code>
      *
+     * @param float $amount Amount value.
      * @param string $currency Currency Code ( GBP, USD, EUR,...)
      * @param array $options Options - "intl", "locale", "symbol",...
      *
      * @return string
-     *
-     * @deprecated since v1.10
      */
-    public function getAmount($currency, $options = array())
+    public static function getAmount($amount, $currency, array $options = array())
     {
-        $useIntl   = ArrayHelper::getValue($options, "intl", false, "bool");
-        $locale    = ArrayHelper::getValue($options, "locale");
-        $symbol    = ArrayHelper::getValue($options, "symbol");
-        $position  = ArrayHelper::getValue($options, "position");
+        $useIntl   = ArrayHelper::getValue($options, 'intl', false, 'bool');
+        $locale    = ArrayHelper::getValue($options, 'locale');
+        $symbol    = ArrayHelper::getValue($options, 'symbol');
+        $position  = ArrayHelper::getValue($options, 'position', 0, 'int');
 
         // Use PHP Intl library.
         if ($useIntl and extension_loaded('intl')) { // Generate currency string using PHP NumberFormatter ( Internationalization Functions )
@@ -119,25 +92,24 @@ class String
             }
 
             $numberFormat = new \NumberFormatter($locale, \NumberFormatter::CURRENCY);
-            $amount       = $numberFormat->formatCurrency($this->content, $currency);
+            $result       = $numberFormat->formatCurrency($amount, $currency);
 
         } else { // Generate a custom currency string.
 
-            if (!empty($symbol)) { // Symbol
+            if (\JString::strlen($symbol) > 0) { // Symbol
 
-                if (0 == $position) { // Symbol at the beginning.
-                    $amount = $symbol . $this->content;
+                if (0 === $position) { // Symbol at the beginning.
+                    $result = $symbol . $amount;
                 } else { // Symbol at end.
-                    $amount = $this->content . $symbol;
+                    $result = $amount . $symbol;
                 }
 
             } else { // Code
-                $amount = $this->content . $currency;
+                $result = $amount . $currency;
             }
-
         }
 
-        return $amount;
+        return $result;
     }
 
     /**
@@ -146,22 +118,18 @@ class String
      * <code>
      * $content = "If you can <strong>dream</strong> it, you can do it. "
      *
-     * $string = new Prism\String($content);
-     * $string->clean();
-     *
-     * echo $string;
+     * echo Prism\Utilities\StringHelper:clean($content);
      * </code>
      *
-     * @return self
-     *
-     * @deprecated since v1.10
+     * @param string $content
+     * @return string
      */
-    public function clean()
+    public static function clean($content)
     {
-        $this->content = strip_tags($this->content);
-        $this->content = \JString::trim(preg_replace('/\r|\n/', ' ', $this->content));
+        $content = strip_tags($content);
+        $content = \JString::trim(preg_replace('/\r|\n/', ' ', $content));
 
-        return $this;
+        return $content;
     }
 
     /**
@@ -172,25 +140,21 @@ class String
      * $length  = 25;
      * $content = "If you can dream it, you can do it."
      *
-     * $string = new Prism\String();
-     * $string->substr($offset, $length);
-     *
-     * echo $string;
+     * echo Prism\Utilities\StringHelper::substr($content, $offset, $length);
      * </code>
      *
+     * @param string $content
      * @param integer $offset
      * @param integer $length
      *
-     * @return self
-     *
-     * @deprecated since v1.10
+     * @return string
      */
-    public function substr($offset, $length)
+    public static function substr($content, $offset, $length)
     {
-        $pos           = \JString::strpos($this->content, ' ', $length);
-        $this->content = \JString::substr($this->content, $offset, $pos);
+        $pos     = \JString::strpos($content, ' ', $length);
+        $content = \JString::substr($content, $offset, $pos);
 
-        return $this;
+        return $content;
     }
 
     /**
@@ -201,30 +165,27 @@ class String
      * // String like this "name1=value1&name2=value2&name3=value3".
      * $rawPost  = file_get_contents("php://input");
      *
-     * $string = new Prism\String($content);
-     * $post   = $string->parseNameValue();
+     * $post = Prism\Utilities\StringHelper::parseNameValue($content);
      * </code>
      *
-     * @return array
+     * @param string $content
      *
-     * @deprecated since v1.10
+     * @return array
      */
-    public function parseNameValue()
+    public static function parseNameValue($content)
     {
         $result = array();
 
-        $data = explode("&", $this->content);
+        $data = explode('&', $content);
         foreach ($data as $value) {
-            $value = explode("=", $value);
+            $value = explode('=', $value);
 
-            if (count($value) == 2) {
-
+            if (count($value) === 2) {
                 $value[0] = urldecode($value[0]);
                 $value[1] = urldecode($value[1]);
 
                 $result[$value[0]] = $value[1];
             }
-
         }
 
         if (!is_array($result)) {
@@ -232,15 +193,5 @@ class String
         }
 
         return $result;
-    }
-
-    /**
-     * Return object value as a string.
-     * 
-     * @return string
-     */
-    public function __toString()
-    {
-        return (string)$this->content;
     }
 }