+++ ui/bootstrap3/fileinput/js/fileinput_locale_cz.js
@@ -21,9 +21,11 @@
         cancelTitle: 'Přerušit  nahrávání',
         uploadLabel: 'Nahrát',
         uploadTitle: 'Nahrát vybrané soubory',
-        msgSizeTooLarge: 'Soubor "{name}" (<b>{size} KB</b>): překročení - maximální povolená velikost <b>{maxSize} KB</b>. Zkuste nahrát znova, prosím!',
-        msgFilesTooLess: 'Musíte vybrat nejméně <b>{n}</b> {files} pro nahrání. Zkuste nahrát znova, prosím!',
-        msgFilesTooMany: 'Počet vybraných souborů pro nahrání <b>({n})</b>: překročení - maximální povolený limit <b>{m}</b>. Zkuste nahrát znova, prosím!',
+        msgZoomTitle: 'zobrazit podrobnosti',
+        msgZoomModalHeading: 'Detailní náhled',
+        msgSizeTooLarge: 'Soubor "{name}" (<b>{size} KB</b>): překročení - maximální povolená velikost <b>{maxSize} KB</b>.',
+        msgFilesTooLess: 'Musíte vybrat nejméně <b>{n}</b> {files} pro nahrání.',
+        msgFilesTooMany: 'Počet vybraných souborů pro nahrání <b>({n})</b>: překročení - maximální povolený limit <b>{m}</b>.',
         msgFileNotFound: 'Soubor "{name}" nebyl nalezen!',
         msgFileSecured: 'Zabezpečení souboru znemožnilo číst soubor "{name}".',
         msgFileNotReadable: 'Soubor "{name}" není čitelný.',
@@ -31,6 +33,7 @@
         msgFilePreviewError: 'Nastala chyba při načtení souboru "{name}".',
         msgInvalidFileType: 'Neplatný typ souboru "{name}". Pouze "{types}" souborů jsou podporovány.',
         msgInvalidFileExtension: 'Neplatná extenze souboru "{name}". Pouze "{extensions}" souborů jsou podporovány.',
+        msgUploadAborted: 'Soubor nahrávání byl přerušen',
         msgValidationError: 'Chyba nahrání souboru.',
         msgLoading: 'Nahrávání souboru {index} z {files} &hellip;',
         msgProgress: 'Nahrávání souboru {index} z {files} - {name} - {percent}% dokončeno.',
@@ -40,6 +43,16 @@
         msgImageHeightSmall: 'Výška image soubor "{name}", musí být alespoň {size} px.',
         msgImageWidthLarge: 'Šířka obrazového souboru "{name}" nelze překročit {size} px.',
         msgImageHeightLarge: 'Výška obrazového souboru "{name}" nelze překročit {size} px.',
-        dropZoneTitle: 'Táhni a pusť soubory sem &hellip;'
+        msgImageResizeError: 'Nelze získat rozměry obrázku změnit velikost.',
+        msgImageResizeException: 'Chyba při změně velikosti obrázku.<pre>{errors}</pre>',
+        dropZoneTitle: 'Táhni a pusť soubory sem &hellip;',
+        fileActionSettings: {
+            removeTitle: 'Odstranit soubor',
+            uploadTitle: 'nahrát soubor',
+            indicatorNewTitle: 'Ještě nenahrál',
+            indicatorSuccessTitle: 'Nahraný',
+            indicatorErrorTitle: 'Nahrát Chyba',
+            indicatorLoadingTitle: 'Nahrávání ...'
+        }
     };
 })(window.jQuery);