+++ ui/bootstrap3/fileinput/js/fileinput_locale_ru.js
@@ -22,9 +22,11 @@
         cancelTitle: 'Отменить текущую загрузку',
         uploadLabel: 'Загрузить',
         uploadTitle: 'Загрузить выбранные файлы',
-        msgSizeTooLarge: 'Файл "{name}" (<b>{size} KB</b>) превышает максимальный размер <b>{maxSize} KB</b>',
-        msgFilesTooLess: 'Вы должны выбрать как минимум <b>{n}</b> {files} для загрузки',
-        msgFilesTooMany: 'Количество выбранных файлов <b>({n})</b> превышает максимально допустимое количество <b>{m}</b>',
+        msgZoomTitle: 'посмотреть детали',
+        msgZoomModalHeading: 'Подробное превью',
+        msgSizeTooLarge: 'Файл "{name}" (<b>{size} KB</b>) превышает максимальный размер <b>{maxSize} KB</b>.',
+        msgFilesTooLess: 'Вы должны выбрать как минимум <b>{n}</b> {files} для загрузки.',
+        msgFilesTooMany: 'Количество выбранных файлов <b>({n})</b> превышает максимально допустимое количество <b>{m}</b>.',
         msgFileNotFound: 'Файл "{name}" не найден!',
         msgFileSecured: 'Ограничения безопасности запрещают читать файл "{name}".',
         msgFileNotReadable: 'Файл "{name}" невозможно прочитать.',
@@ -32,6 +34,7 @@
         msgFilePreviewError: 'Произошла ошибка при чтении файла "{name}".',
         msgInvalidFileType: 'Запрещенный тип файла для "{name}". Только "{types}" разрешены.',
         msgInvalidFileExtension: 'Запрещенное расширение для файла "{name}". Только "{extensions}" разрешены.',
+        msgUploadAborted: 'Выгрузка файла прервана',
         msgValidationError: 'Ошибка при загрузке файла',
         msgLoading: 'Загрузка файла {index} из {files} &hellip;',
         msgProgress: 'Загрузка файла {index} из {files} - {name} - {percent}% завершено.',
@@ -41,6 +44,16 @@
         msgImageHeightSmall: 'Высота изображения {name} должна быть не меньше {size} px.',
         msgImageWidthLarge: 'Ширина изображения "{name}" не может превышать {size} px.',
         msgImageHeightLarge: 'Высота изображения "{name}" не может превышать {size} px.',
-        dropZoneTitle: 'Перетащите файлы сюда &hellip;'
+        msgImageResizeError: 'Не удалось получить размеры изображения, чтобы изменить размер.',
+        msgImageResizeException: 'Ошибка при изменении размера изображения.<pre>{errors}</pre>',
+        dropZoneTitle: 'Перетащите файлы сюда &hellip;',
+        fileActionSettings: {
+            removeTitle: 'Удалить файл',
+            uploadTitle: 'Загрузить файл',
+            indicatorNewTitle: 'Еще не загружен',
+            indicatorSuccessTitle: 'Загружен',
+            indicatorErrorTitle: 'Ошибка загрузки',
+            indicatorLoadingTitle: 'Загрузка ...'
+        }
     };
 })(window.jQuery);