+++ ui/bootstrap3/fileinput/js/fileinput_locale_tr.js
@@ -21,9 +21,11 @@
         cancelTitle: 'Devam eden yüklemeyi iptal et',
         uploadLabel: 'Yükle',
         uploadTitle: 'Seçilen dosyaları yükle',
-        msgSizeTooLarge: '"{name}" dosyasının boyutu (<b>{size} KB</b>) izin verilen azami dosya boyutu olan <b>{maxSize} KB</b>\'tan büyük. Lütfen tekrar deneyin!',
-        msgFilesTooLess: 'Yüklemek için en az <b>{n}</b> {files} dosya seçmelisiniz. Lütfen tekrar deneyin!',
-        msgFilesTooMany: 'Yüklemek için seçtiğiniz dosya sayısı <b>({n})</b> azami limitin <b>{m}</b> altında olmalıdır. Lütfen tekrar deneyin!',
+        msgZoomTitle: 'Ayrıntıları görüntüle',
+        msgZoomModalHeading: 'Detaylı Önizleme',
+        msgSizeTooLarge: '"{name}" dosyasının boyutu (<b>{size} KB</b>) izin verilen azami dosya boyutu olan <b>{maxSize} KB</b>\'tan büyük.',
+        msgFilesTooLess: 'Yüklemek için en az <b>{n}</b> {files} dosya seçmelisiniz.',
+        msgFilesTooMany: 'Yüklemek için seçtiğiniz dosya sayısı <b>({n})</b> azami limitin <b>{m}</b> altında olmalıdır.',
         msgFileNotFound: '"{name}" dosyası bulunamadı!',
         msgFileSecured: 'Güvenlik kısıtlamaları "{name}" dosyasının okunmasını engelliyor.',
         msgFileNotReadable: '"{name}" dosyası okunabilir değil.',
@@ -31,6 +33,7 @@
         msgFilePreviewError: '"{name}" dosyası okunurken bir hata oluştu.',
         msgInvalidFileType: '"{name}" dosyasının türü geçerli değil. Yalnızca "{types}" türünde dosyalara izin veriliyor.',
         msgInvalidFileExtension: '"{name}" dosyasının uzantısı geçersiz. Yalnızca "{extensions}" uzantılı dosyalara izin veriliyor.',
+        msgUploadAborted: 'Dosya yükleme iptal edildi',
         msgValidationError: 'Dosya Yükleme Hatası',
         msgLoading: 'Dosyalar yükleniyor {index} / {files} &hellip;',
         msgProgress: 'Dosya yükleniyor {index} / {files} - {name} - %{percent} tamamlandı.',
@@ -40,6 +43,16 @@
         msgImageHeightSmall: 'Görüntü dosyası "{name}" yüksekliği en az {size} piksel olmalıdır.',
         msgImageWidthLarge: 'Görüntü dosyası "{name}" genişliği {size} px geçemez.',
         msgImageHeightLarge: 'Resim dosyası "{name}" Yükseklik {size} px geçemez.',
-        dropZoneTitle: 'Dosyaları buraya sürükleyip bırakın &hellip;'
+        msgImageResizeError: 'Görüntü boyutlarını yeniden boyutlandırmak için alınamadı.',
+        msgImageResizeException: 'Hata resim boyutlandırma sırasında.<pre>{errors}</pre>',
+        dropZoneTitle: 'Dosyaları buraya sürükleyip bırakın &hellip;',
+        fileActionSettings: {
+            removeTitle: 'dosyayı kaldır',
+            uploadTitle: 'dosya yükleme',
+            indicatorNewTitle: 'Henüz yüklendi',
+            indicatorSuccessTitle: 'yüklendi',
+            indicatorErrorTitle: 'Yükleme Hatası',
+            indicatorLoadingTitle: 'Yükleme ...'
+        }
     };
 })(window.jQuery);