+++ ui/bootstrap3/fileinput/js/fileinput_locale_uk.js
@@ -22,9 +22,11 @@
         cancelTitle: 'Скасувати поточну загрузку',
         uploadLabel: 'Загрузити',
         uploadTitle: 'Загрузити вибрані файли',
-        msgSizeTooLarge: 'Файл "{name}" (<b>{size} KB</b>) перевищує максимальний розмір <b>{maxSize} KB</b>',
-        msgFilesTooLess: 'Ви повинні вибрати як мінімум <b>{n}</b> {files} для загрузки',
-        msgFilesTooMany: 'Кількість вибраних файлів <b>({n})</b> перевищує максимально допустиму кількість <b>{m}</b>',
+        msgZoomTitle: 'Подивитися деталі',
+        msgZoomModalHeading: 'Детальний превью',
+        msgSizeTooLarge: 'Файл "{name}" (<b>{size} KB</b>) перевищує максимальний розмір <b>{maxSize} KB</b>.',
+        msgFilesTooLess: 'Ви повинні вибрати як мінімум <b>{n}</b> {files} для загрузки.',
+        msgFilesTooMany: 'Кількість вибраних файлів <b>({n})</b> перевищує максимально допустиму кількість <b>{m}</b>.',
         msgFileNotFound: 'Файл "{name}" не знайдено!',
         msgFileSecured: 'Обмеження безпеки перешкоджають читанню файла "{name}".',
         msgFileNotReadable: 'Файл "{name}" неможливо прочитати.',
@@ -32,6 +34,7 @@
         msgFilePreviewError: 'Сталася помилка під час читання файла "{name}".',
         msgInvalidFileType: 'Заборонений тип файла для "{name}". Тільки "{types}" дозволені.',
         msgInvalidFileExtension: 'Заборонене розширення для файла "{name}". Тільки "{extensions}" дозволені.',
+        msgUploadAborted: 'Вивантаження файлу перервана',
         msgValidationError: 'Помилка під час загрузки файла',
         msgLoading: 'Загрузка файла {index} із {files} &hellip;',
         msgProgress: 'Загрузка файла {index} із {files} - {name} - {percent}% завершено.',
@@ -41,6 +44,16 @@
         msgImageHeightSmall: 'Висота зображення "{name}" повинна бути не менше {size} px.',
         msgImageWidthLarge: 'Ширина зображення "{name}" не може перевищувати {size} px.',
         msgImageHeightLarge: 'Висота зображення "{name}" не може перевищувати {size} px.',
-        dropZoneTitle: 'Перетягніть файли сюди &hellip;'
+        msgImageResizeError: 'Не вдалося розміри зображення, щоб змінити розмір.',
+        msgImageResizeException: 'Помилка при зміні розміру зображення.<pre>{errors}</pre>',
+        dropZoneTitle: 'Перетягніть файли сюди &hellip;',
+        fileActionSettings: {
+            removeTitle: 'Видалити файл',
+            uploadTitle: 'Загрузити файл',
+            indicatorNewTitle: 'Ще не загружено',
+            indicatorSuccessTitle: 'Загружено',
+            indicatorErrorTitle: 'Помилка при загрузці',
+            indicatorLoadingTitle: 'Загрузка ...'
+        }
     };
 })(window.jQuery);