+++ ui/layouts/bootstrap3/endtabset.php
@@ -8,7 +8,5 @@
  */
 
 defined('_JEXEC') or die;
-
 ?>
-
 </div>